Restructure of the app
