+++ app/routes/listing.py
@@ -1,11 +1,10 @@
 from flask import Blueprint, request, session, redirect, abort, render_template, url_for, jsonify, Response, flash
-from flask_bcrypt import Bcrypt
-from data.models import Users, Listing, db
+from data.models import Users, Listing
 from werkzeug.utils import secure_filename
+from utils.helpers import db,bcrypt
 
 # Create blueprint for authentication related routes
 listing_bp = Blueprint('listing', __name__)
-bcrypt = Bcrypt()
 
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
 ALLOWED_MIME_TYPES = {'image/png', 'image/jpeg', 'image/gif'}
