+++ app/static/fonts/gothic.otf
